demrepofdave/1-fix-compiler-warnings: Fixed possible memory overrun issues with directory and filenames. Removed magic numbers.  Resized all directory/filenames buffers from 512 to 768 as modern OS's can exceed 512 bytes for filepath + filename.

+++ src/linux-gui.c
@@ -506,7 +506,7 @@ MENU dischmenu[4]=
 
 int gui_mrbmode()
 {
-        mrbmode=(int)active_menu->dp;
+        mrbmode=(intptr_t)active_menu->dp;
         resetit=1;
         updatelinuxgui();
         return D_O_K;
